refactor(loading-inventories): Concurrent inventory loading.
Now the loading of both owned and shared inventories is concurrent, that
is, they will be loaded at the same time and won't block each other, so
the longest wait time will be whichever takes longer and not the sum of
both wait times. Also I refactored the code a bit to make it look better
and to also use a more functional approach rather than what was being
done.